+++ src/hyperfinehamiltonian.jl
@@ -8,8 +8,8 @@ This generates a list of all J₃ values that satisfy the Triangle Conditions wi
    inputs J₁ and J₂. This is typically used to determine N values for a given
    J and S.
 """
-   max = Int64(J+S)
-   min = Int64(abs(J-S))
+   max = Int(J+S)
+   min = Int(abs(J-S))
    return collect(min:max)
 end
 function srprep(J,S)
